// TrsFrontend/src/components/TrsCalculation/Interfaces/StyleInfo.tsx
export interface StyleInfo {
  image: string;
  authorId: number;
  buyerName: string;
  created: Date;
  editorId: number;
  generalAllowance: number;
  id: number;
  modified: Date;
  season: string;
  size: string;
  styleCode: string;
  styleDescription: string;
  styleName: string;
  styleNumber: string;
  thicknessFabric: number;
  totalGarments: number;
 
}